Fix float literals narrowed to single precision in parser
parse_primary cast every FLOAT_LITERAL through static_cast<float>, so a
DOUBLE literal like 3.141592653589793 was permanently degraded to
3.1415927. It then no longer compared equal to the same value stored in
a DOUBLE column, and inserts into DOUBLE columns silently lost precision.

Parse literals as double (the wider type); insertion into a FLOAT column
still narrows via the schema-driven cast. Add a sql_regression case
asserting a DOUBLE literal round-trips and compares equal, while a value
differing only beyond float precision does not match.

+++ src/sql/parser/parser.cpp
@@ -1181,7 +1181,12 @@ UniquePtr<Expression> Parser::parse_primary() {
         lexer_.consume_token();
         auto expr = make_unique<Expression>();
         expr->type = ExprType::kLiteral;
-        expr->literal_value = Value(static_cast<float>(atof(t.value.c_str())));
+        // Parse as double (the wider type). Narrowing to float here would
+        // silently drop precision for every literal — e.g. the DOUBLE literal
+        // 3.141592653589793 became 3.1415927, so it no longer compared equal
+        // to the same value stored in a DOUBLE column. Insertion into a FLOAT
+        // column still narrows correctly via the schema-driven cast.
+        expr->literal_value = Value(static_cast<double>(atof(t.value.c_str())));
         return expr;
     }
